+++ src/voxels/voxel.h
@@ -33,20 +33,20 @@ private:
     bool _visible;
 public:
     Voxel();
-    Voxel(glm::vec3 position);
-    Voxel(glm::vec3 position, glm::vec4 color);
-    Voxel(glm::vec3 position, glm::vec3 velocity, glm::vec4 color);
+    Voxel(glm::ivec3 position);
+    Voxel(glm::ivec3 position, glm::vec4 color);
+    Voxel(glm::ivec3 position, glm::vec3 velocity, glm::vec4 color);
 
     ~Voxel();
 
-    void setPosition(glm::vec3 position);
+    void setPosition(glm::ivec3 position);
 
     /**
      * @brief Get the Position object
      * 
      * @return glm::vec3 
      */
-    glm::vec3 getPosition();
+    glm::ivec3 getPosition();
 
     void setVelocity(glm::vec3 velocity);
     glm::vec3 getVelocity();
